ejercicio-2/backend/requirements.txt
alembic==1.12.0
APIFlask==2.0.2
apispec==6.3.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mysqlclient==2.2.0
packaging==23.2
python-dotenv==0.21.1
SQLAlchemy==2.0.22
typing_extensions==4.8.0
webargs==8.3.0
Werkzeug==3.0.0
